+++ native/core/src/execution/datafusion/expressions/mod.rs
@@ -43,10 +43,10 @@ mod utils;
 pub mod variance;
 pub mod xxhash64;
 
-pub use datafusion_comet_spark_expr::EvalMode;
+pub use datafusion_comet_spark_expr::{EvalMode, SparkError};
 
 fn arithmetic_overflow_error(from_type: &str) -> CometError {
-    CometError::ArithmeticOverflow {
+    CometError::Spark(SparkError::ArithmeticOverflow {
         from_type: from_type.to_string(),
-    }
+    })
 }
